backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
openai==2.14.0
python-dotenv==1.0.1
requests==2.31.0
pywebpush==1.14.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
